Move component regestry to component package so it can be reused outside of runtime package
