+++ internal/runtime/internal/controller/component_node_manager.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"sync"
 
+	"github.com/grafana/alloy/internal/component"
 	"github.com/grafana/alloy/syntax/ast"
 )
 
@@ -12,7 +13,7 @@ import (
 type ComponentNodeManager struct {
 	globals ComponentGlobals
 	// builtinComponentReg returns information to build and run built-in components.
-	builtinComponentReg ComponentRegistry
+	builtinComponentReg component.ComponentRegistry
 
 	mut sync.RWMutex
 	// customComponentReg returns information to build and run custom components.
@@ -20,7 +21,7 @@ type ComponentNodeManager struct {
 }
 
 // NewComponentNodeManager creates a new ComponentNodeManager without custom component registry.
-func NewComponentNodeManager(globals ComponentGlobals, componentReg ComponentRegistry) *ComponentNodeManager {
+func NewComponentNodeManager(globals ComponentGlobals, componentReg component.ComponentRegistry) *ComponentNodeManager {
 	return &ComponentNodeManager{
 		globals:             globals,
 		builtinComponentReg: componentReg,
